refactor: extract ligand modal
